Suppressed Unnecessary Warnings

+++ src/test/java/DataStructures/MinPriorityQueue/MinHeapTest.java
@@ -1,10 +1,13 @@
 package DataStructures.MinPriorityQueue;
 
-import static org.junit.Assert.*;
-
 import datastructures.minpriorityqueue.MinHeap.MinHeap;
-import java.util.*;
 import org.junit.Test;
+
+import java.util.Arrays;
+import java.util.HashSet;
+import java.util.PriorityQueue;
+
+import static org.junit.Assert.*;
 
 public class MinHeapTest {
 
@@ -53,7 +56,8 @@ public class MinHeapTest {
         }
       }
 
-      assertEquals((int) priorityQueue.poll(), (int) heap.extractMin());
+      //noinspection ConstantConditions
+      assertEquals((int) priorityQueue.poll(), heap.extractMin());
     }
   }
